rcu: Add READ_ONCE() to rcu_do_batch() access to rcu_resched_ns
Given that sysfs can change the value of rcu_resched_ns at any time,
this commit adds a READ_ONCE() to the sole access to that variable.
While in the area, this commit also adds bounds checking, clamping the
value to at least a millisecond, but no longer than a second.

Change-Id: I1e5f126a173d51c3b8f40d2d86360bf589dcbb14
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>